Changes to the math parser that are necissary but broke some things.
I switched the math parser to phrase mode.  This means that it won't
bail out with a success if it can parse half way through a line but
fails somewhere else.  Before, something like this would work:

    1 + 2 / 3,

even though the comma is illegal. (in this case it would parse
1 + 2 / 3 correctly and ignore the rest)

with phrase mode, this is illegal, because of this, we are failing
some tests.